made a class for methods
