+++ Factorial/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\222257\\AppData\\Roaming\\NetBeans\\11.3\\build.properties
+user.properties.file=C:\\Users\\222257\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
